# backend/src/openapi/wire.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WIRE API",
    "version": "1.0.0",
    "description": "REST API for wire transfer intent management, approvals, voice challenges, audit bundles, API keys, and webhooks.\n\nThis API provides:\n- Transfer intent lifecycle (create, update, execute)\n- Voice challenge and proof submission\n- Approval decisions and token management\n- Beneficiary management\n- Policy configuration and simulation\n- Event chain and audit bundle generation\n- API keys and webhooks (developer platform)\n\nAll endpoints are prefixed with `/api/wire`.\n\n**Authentication:**\n- UI sessions: `Authorization: Bearer <token>`\n- Integrations: `Authorization: Basic base64(<apiKey>:<apiSecret>)` (recommended) or `X-API-KEY: <apiKey>` (legacy)\n\n**Error Format:** All errors follow `{ error: string, code?: string, details?: object }` format.\n**Idempotency:** Mutating endpoints support `X-Idempotency-Key` (where applicable)."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/wire",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/beneficiaries": {
      "get": {
        "tags": ["Beneficiaries"],
        "summary": "List beneficiaries",
        "description": "Returns all beneficiaries for the authenticated user's organization.",
        "operationId": "listBeneficiaries",
        "responses": {
          "200": {
            "description": "List of beneficiaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Beneficiary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Beneficiaries"],
        "summary": "Create beneficiary",
        "description": "Creates a new beneficiary. Requires `beneficiary:create` permission.",
        "operationId": "createBeneficiary",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBeneficiaryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Beneficiary created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Beneficiary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/beneficiaries/{id}": {
      "patch": {
        "tags": ["Beneficiaries"],
        "summary": "Update beneficiary",
        "description": "Updates a beneficiary. Requires `beneficiary:create` permission.",
        "operationId": "updateBeneficiary",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "description": "Beneficiary ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBeneficiaryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Beneficiary updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Beneficiary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Beneficiary not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "UI session auth (Bearer token)."
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Integration auth: base64(<apiKey>:<apiSecret>)"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "Legacy integration auth (key-only). Prefer Basic auth."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-parseable error code (optional)"
          },
          "details": {
            "type": "object",
            "description": "Additional error details (optional)"
          }
        }
      },
...
      "Beneficiary": {
        "type": "object",
        "required": [
          "id",
          "orgId",
          "displayName",
          "country",
          "railsAllowed",
          "bankLast4",
          "bankTokenHash",
          "version",
          "status",
          "createdAt",
          "updatedAt",
          "lastChangedAt",
          "lastChangedBy"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Beneficiary ID"
          },
          "orgId": {
            "type": "string",
            "description": "Organization ID"
          },
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "description": "ISO 3166-1 alpha-2 country code"
          },
          "railsAllowed": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ACH", "WIRE"]
            },
            "minItems": 1,
            "description": "Allowed payment rails"
          },
          "bankLast4": {
            "type": "string",
            "minLength": 4,
            "maxLength": 4,
            "description": "Last 4 digits of bank account"
          },
          "bankTokenHash": {
            "type": "string",
            "minLength": 1,
            "description": "Hash of encrypted bank token"
          },
          "version": {
            "type": "integer",
            "description": "Version number"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "LOCKED", "PENDING_VERIFICATION"],
            "description": "Beneficiary status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "lastChangedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last change timestamp"
          },
          "lastChangedBy": {
            "type": "string",
            "description": "User ID who last changed the beneficiary"
          }
        }
      },
      "CreateBeneficiaryRequest": {
        "type": "object",
        "required": [
          "displayName",
          "country",
          "railsAllowed",
          "bankLast4",
          "bankTokenHash"
        ],
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "description": "Display name"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "description": "ISO 3166-1 alpha-2 country code"
          },
          "railsAllowed": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ACH", "WIRE"]
            },
            "minItems": 1,
            "description": "Allowed payment rails"
          },
          "bankLast4": {
            "type": "string",
            "minLength": 4,
            "maxLength": 4,
            "description": "Last 4 digits of bank account"
          },
          "bankTokenHash": {
            "type": "string",
            "minLength": 1,
            "description": "Hash of encrypted bank token"
          }
        }
      },
      "UpdateBeneficiaryRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "description": "Display name"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "description": "ISO 3166-1 alpha-2 country code"
          },
          "railsAllowed": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ACH", "WIRE"]
            },
            "minItems": 1,
            "description": "Allowed payment rails"
          },
          "bankLast4": {
            "type": "string",
            "minLength": 4,
            "maxLength": 4,
            "description": "Last 4 digits of bank account"
          },
          "bankTokenHash": {
            "type": "string",
            "minLength": 1,
            "description": "Hash of encrypted bank token"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "LOCKED", "PENDING_VERIFICATION"],
            "description": "Beneficiary status"
          }
        },
        "additionalProperties": false,
        "minProperties": 1,
        "description": "At least one field must be provided"
      },
...
    }
  }
}